+++ clang-tools-extra/clang-doc/markdown/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(LLVM_LINK_COMPONENTS
+  Support
+  )
+
+add_clang_library(clangDocMarkdown STATIC
+  Markdown.cpp
+  )
